feat: add protected layout and auth middleware
