+++ app/(protected)/app/home/page.tsx
@@ -1,15 +1,10 @@
 'use client';
 
 import { useState } from 'react';
-import { MainLayout } from '@/layouts/MainLayout';
 import { Home as HomeView } from '@/views/app/Home';
 
 export default function Page() {
   const [theme, setTheme] = useState<'ios' | 'material'>('material');
 
-  return (
-    <MainLayout>
-      <HomeView theme={theme} onTheme={setTheme} />
-    </MainLayout>
-  );
+  return <HomeView theme={theme} onTheme={setTheme} />;
 }
