+++ app/(protected)/app/icons/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { IconsDemo } from '@/views/app/IconsDemo';
+
+export default function Page() {
+  return <IconsDemo />;
+}
