--- app/app/icons/page.tsx
@@ -1,12 +1,0 @@
-'use client';
-
-import { MainLayout } from '@/layouts/MainLayout';
-import { IconsDemo } from '@/views/app/IconsDemo';
-
-export default function Page() {
-  return (
-    <MainLayout>
-      <IconsDemo />
-    </MainLayout>
-  );
-}
