+++ middleware.ts
@@ -1,0 +1,9 @@
+import { authMiddleware } from '@clerk/nextjs';
+
+export default authMiddleware({
+  publicRoutes: ['/', '/auth/(.*)'],
+});
+
+export const config = {
+  matcher: ['/((?!.*\\..*|_next).*)'],
+};
